Created overlays directory for overlay components. Moved existing overlay components to overlays directory, updated relevant reference paths.
