+++ src/frontend/utilities/overlays/WeatherOverlay.jsx
